Use only one MediaPlayer object
Previously, 2 media players were used using setNextMediaPlayer(). The idea
is that whenever one finishes playing a song, the second starts playing the
next song with as little delay as possible.

However, due to unknown reasons, this causes the following error if the
media players are reinitialized 16 or more times (e.g. when selecting a
folder to play 16 times in a row):

E/MediaPlayer: error (1, -19)

After this, only a phone reboot can allow playing more music.

There is no authoritative information about this error, but people say it's
because too many a media players are created and not release.
However, tracing with a debugger shows that for every new media player
created, the previous one is release()d.

The solution is going back to a single media player and accepting the small
delay between files. I don't notice this delay, but perhaps it's there.

With this change, the error (1, -19) does not appear anymore, however many
times a media player is reloaded.