Workaround ANR on bluetooth connect
When the user has quit SMP and then connects to Bluetooth which sends a
PLAY event, this event is routed to SMPMediaButtonReceiver even though
we unregistered the even listener on shutdown. SMPMediaButtonReceiver
then starts the service, which fails since the app is in the background.

The workaround is to catch the exception and print a log message.